--- About.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-function About()
-{
-    return (
-        <h1>About Page</h1>
-    )
-}
-
-export default About;
