--- Error.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-function Error()
-{
-    return (
-        <h1>404 Page Not Found</h1>
-    )
-}
-
-export default Error;
