--- Facts.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-function Facts()
-{
-    return (
-        <h1>Facts</h1>
-    )
-}
-
-export default Facts;
