--- Gallery.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-function Gallery()
-{
-    return (
-        <h1>Gallery</h1>
-    )
-}
-
-export default Gallery;
