--- index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <link rel="stylesheet" href="style.css">
-    </head>
-    <body>
-        <div id="navbar"></div>
-        <div id="page"></div>
-        <script src="Selection.pack.js"></script>
-    </body>
-</html>
