+++ src/components/Facts.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+function Facts()
+{
+    return (
+        <h1>Facts</h1>
+    )
+}
+
+export default Facts;
