+++ src/components/Gallery.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+function Gallery()
+{
+    return (
+        <h1>Gallery</h1>
+    )
+}
+
+export default Gallery;
